+++ test/runtests.jl
@@ -10,6 +10,10 @@ import Wilsonloop: loops_staple
     include("high_level_api.jl")
 end
 
+@testset "Normalized HYP smearing" begin
+    include("nhyp_smearing.jl")
+end
+
 @testset "Molecular dynamics driver" begin
     include("md_driver.jl")
 end